:bug: Fix handling of multiple roots per repo
This fixes the issue that in case there are multiple project roots for a
single repo only the root which matches exactly the repository's working
directory was updated. Also it fixes the issue that an uncaught type
error was thrown in case a sub-folder of a repository has been selected
as project root.

closes #1