Added options for Git Flow support
This adds the options for git flow to the settings menu, and orders all items
with a fixed position. The git flow options have their own settings group,
since they belong together and have their own on-off switch.

Also adds the listeners to the main.coffee and treeviewui.coffee files